Added wpf

+++ Diplom/CallAPI/Brigade/AssignCallHandler.cs
@@ -21,7 +21,7 @@ namespace CallAPI.Brigade
         {
             var call = await _databaseProvider.InDatabaseScope(context => context.Calls.FirstAsync(e => e.Id == request.Request.CallId), cancellationToken);
             call.AmbulanceBrigadeId = request.Request.BrigadeId;
-            call.Status = CallStatus.TransferedToBrigade;
+            call.Status = (byte)CallStatus.TransferedToBrigade;
             _ = await _databaseProvider.InDatabaseScope(context => context.Calls.Update(call), cancellationToken);
             return Guid.NewGuid();
         }
